create auth in store folder (WIP)
