+++ src/app/store/user/selectors.ts
@@ -1,5 +1,6 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { UserState, DashboardState } from './reducers';
+import { state } from "@angular/animations";
 
 export const userFeature = createFeatureSelector<UserState>('userData');
 export const dashboardFeature = createFeatureSelector<DashboardState>('dashboard');
@@ -17,6 +18,11 @@ export const selectAllUsers = createSelector(
 export const selectAssessmentsData = createSelector(
   dashboardFeature,
   (state) => state.assessments,
+);
+
+export const selectAssessmentsSize = createSelector(
+  dashboardFeature,
+  (state) => state.assessmentSize,
 );
 
 export const selectAssessmentReport = createSelector(
